Harden the init disk imaging functions to deal with weird disks
Either the disks may not be aligned to the MB, or that the drive
has more to read than is reported, either way the old function
would produce files that would be un-restorable since the header would
have already been emitted incorrecly, since hot-clone is pipeable
this is unfixable.

Instead we ensure we only read what we said we would, and then
we try and read more to see if the deive is broken and/or mis
reporting itself.